feat: support tf2.6.1 (#45)

+++ musa_ext/kernels/array/musa_strided_slice_op.cc
@@ -33,7 +33,7 @@ class MusaStridedSliceOp : public OpKernel {
 
     PartialTensorShape processing_shape, final_shape;
     bool is_identity = true, slice_dim0 = true, is_simple_slice = true;
-    gtl::InlinedVector<long long, 4> begin, end, strides;
+    gtl::InlinedVector<int64_t, 4> begin, end, strides;
 
     OP_REQUIRES_OK(
         context, ::tensorflow::ValidateStridedSliceOp(
